Add Timer module and some changes regards to pull request comments
